+++ src/index.ts
@@ -1,1 +1,2 @@
+export type { RerunWithoutBundlingEvent } from "./domain/RerunWithoutBundlingEvent";
 export { ServerProxy } from "./application/ServerProxy";